Final version 100/100
Got 100/100 on testing system
Edited files with code style rules
