+++ 3-Petroshenko-B15/.gitignore
@@ -1,0 +1,8 @@
+**Debug
+**Release
+**.vs
+**.sln
+**.vcxproj
+**.vcxproj.filters
+**.vcxproj.user
+**.config
