+++ 3-Petroshenko-B15/LabB-Var15/LabB.c
@@ -1,17 +1,10 @@
-#include <stdio.h>
-#include <stdlib.h>
-#include <memory.h>
-#include <math.h>
-#include "memallocator.h"
+#include "LabB.h"
+#pragma warning(disable:4996)
 
-typedef enum {
-	NoMerge,
-	MergeRight,
-	MergeLeft
-}Merge;
-
-#define Ok 1
-#define Error 0
+#define TEST_BLOCKS_COUNT 10000
+#define ARR_SIZE 1024
+#define BLOCK_MAX 16
+#define RUN_COUNT 25
 
 //pointer on the begining of the using heap
 void* sMemory = NULL;
@@ -38,10 +31,10 @@ int* GetFlagPtr(void* descriptor) {
 	return (int*)((char*)descriptor + *GetSizePtr(descriptor) - sizeof(int));
 }
 
-int meminit(void* pMemory, int size) {
+int MemInit(void* pMemory, int size) {
 	if (!pMemory)
 		return Error;
-	if (size <= memgetminimumsize())
+	if (size <= MemGetMinimumSize())
 		return Error;
 	void* initdesc = pMemory;
 	*GetSizePtr(initdesc) = size;
@@ -54,7 +47,7 @@ int meminit(void* pMemory, int size) {
 	return Ok;
 }
 
-void memdone() {
+void MemDone() {
 	void* iterator = sMemory;
 	while ((char*)iterator < (char*)sMemory + sizeofs)
 	{
@@ -66,10 +59,10 @@ void memdone() {
 	return;
 }
 
-void* memalloc(int size) {
+void* MemAlloc(int size) {
 	if (size <= 0)
 		return NULL;
-	if (size > sizeofs - memgetblocksize())
+	if (size > sizeofs - MemGetBlockSize())
 		return NULL;
 	void* iterator = sHead;
 	void* fitblock = NULL;
@@ -77,7 +70,7 @@ void* memalloc(int size) {
 	//searching for best fit block
 	while (iterator != NULL)
 	{
-		if (*GetSizePtr(iterator) >= size + memgetblocksize() && *GetSizePtr(iterator) < min)
+		if (*GetSizePtr(iterator) >= size + MemGetBlockSize() && *GetSizePtr(iterator) < min)
 		{
 			fitblock = iterator;
 			min = *GetSizePtr(iterator);
@@ -88,11 +81,11 @@ void* memalloc(int size) {
 		return NULL;
 	//fragmentation:
 	//if we have enough memory to create new descriptor for empty part
-	if (*GetSizePtr(fitblock) > memgetblocksize() + size + memgetblocksize())
+	if (*GetSizePtr(fitblock) > MemGetBlockSize() + size + MemGetBlockSize())
 	{
 		//right part of the block, i.e. empty part
-		void* newdescriptor = (void*)((char*)fitblock + memgetblocksize() + size);
-		*GetSizePtr(newdescriptor) = *GetSizePtr(fitblock) - size - memgetblocksize();
+		void* newdescriptor = (void*)((char*)fitblock + MemGetBlockSize() + size);
+		*GetSizePtr(newdescriptor) = *GetSizePtr(fitblock) - size - MemGetBlockSize();
 		*GetNextPtr(newdescriptor) = *GetNextPtr(fitblock);
 		*GetPrevPtr(newdescriptor) = *GetPrevPtr(fitblock);
 		*GetFlagPtr(newdescriptor) = -*GetSizePtr(newdescriptor);
@@ -111,7 +104,7 @@ void* memalloc(int size) {
 				*GetPrevPtr(*GetNextPtr(fitblock)) = newdescriptor;
 			*GetNextPtr(*GetPrevPtr(fitblock)) = newdescriptor;
 		}
-		*GetSizePtr(fitblock) = memgetblocksize() + size;
+		*GetSizePtr(fitblock) = MemGetBlockSize() + size;
 	}
 	//if we give all block to user
 	else
@@ -119,27 +112,24 @@ void* memalloc(int size) {
 		//if this block is the first
 		if (!*GetPrevPtr(fitblock))
 		{
-			//if this block is not the only one in the list of empty blocks
-			if (*GetNextPtr(fitblock))
+			//if this block is the only one in the list of empty blocks
+			if (!*GetNextPtr(fitblock))
+				sHead = NULL;
+			else
 			{
 				*GetPrevPtr(*GetNextPtr(fitblock)) = NULL;
 				sHead = *GetNextPtr(fitblock);
 			}
-			else
-				sHead = NULL;
 		}
 		else
 		{
-			//if this block is not the last
-			if (*GetNextPtr(fitblock))
+			//if this block is the last
+			if (!*GetNextPtr(fitblock))
+				*GetNextPtr(*GetPrevPtr(fitblock)) = NULL;
+			else
 			{
 				*GetPrevPtr(*GetNextPtr(fitblock)) = *GetPrevPtr(fitblock);
 				*GetNextPtr(*GetPrevPtr(fitblock)) = *GetNextPtr(fitblock);
-			}
-			else
-			{
-				*GetNextPtr(*GetPrevPtr(fitblock)) = NULL;
-				*GetPrevPtr(*GetNextPtr(fitblock)) = *GetPrevPtr(fitblock);
 			}
 		}
 	}
@@ -151,7 +141,7 @@ void* memalloc(int size) {
 	return (void*)((char*)fitblock + sizeof(int) + sizeof(void*) + sizeof(void*));
 }
 
-void memfree(void* p) {
+void MemFree(void* p) {
 	if (!p)
 		return;
 	void* pdescriptor = (void*)((char*)p - sizeof(int) - sizeof(void*) - sizeof(void*));
@@ -187,10 +177,21 @@ void memfree(void* p) {
 			sHead = pdescriptor;
 		}
 		//repoint next and prev pointers to each other not to lose them after merge
-		if (*GetPrevPtr(rightblock))
-			*GetNextPtr(*GetPrevPtr(rightblock)) = *GetNextPtr(rightblock);
-		if (*GetNextPtr(rightblock))
-			*GetPrevPtr(*GetNextPtr(rightblock)) = *GetPrevPtr(rightblock);
+		//if right block is the first
+		if (!*GetPrevPtr(rightblock))
+		{
+			//if right block is not the only one
+			if (*GetNextPtr(rightblock))
+				*GetPrevPtr(*GetNextPtr(rightblock)) = *GetPrevPtr(rightblock); // = NULL
+			sHead = *GetNextPtr(rightblock);
+		}
+		else
+		{
+			//if right block is not the last
+			if (*GetNextPtr(rightblock))
+				*GetPrevPtr(*GetNextPtr(rightblock)) = *GetPrevPtr(rightblock);
+			*GetNextPtr(*GetPrevPtr(rightblock)) = *GetNextPtr(rightblock);// maybe NULL
+		}
 		*GetSizePtr(pdescriptor) += *GetSizePtr(rightblock);
 		*GetFlagPtr(pdescriptor) = -*GetSizePtr(pdescriptor);
 	}
@@ -207,10 +208,93 @@ void memfree(void* p) {
 	return;
 }
 
-int memgetminimumsize() {
+int MemGetMinimumSize() {
 	return sizeof(int) + sizeof(void*) + sizeof(void*) + sizeof(int);
 }
 
-int memgetblocksize() {
+int MemGetBlockSize() {
 	return sizeof(int) + sizeof(void*) + sizeof(void*) + sizeof(int);
 }
+
+int main(int argc, char* argv[]) {
+	/*int seed;
+	if (argc > 1) {
+		sscanf(argv[1], "%d", &seed);
+	}
+	else {
+		srand(time(0));
+		seed = rand() % 1024;
+		printf("seed: %d\n", seed);
+	}*/
+	srand(180);
+	void* arr[ARR_SIZE];
+	char* ones[BLOCK_MAX];
+	int sizes[ARR_SIZE];
+	memset(ones, '1', BLOCK_MAX * sizeof(ones[0]));
+	void* buf = malloc(1000000);
+	MemInit(buf, 1000000);
+	for (int r = 0; r < RUN_COUNT; r++) {
+		for (int i = 0; i < ARR_SIZE; i++) {
+			int size = 1 + rand() % BLOCK_MAX;
+			arr[i] = MemAlloc(size);
+			sizes[i] = size;
+			memset(arr[i], '1', size);
+		}
+		for (int i = 0; i < ARR_SIZE; i++) {
+			int index = rand() % (ARR_SIZE - i);
+			if (memcmp(arr[index], ones, sizes[index])) {
+				printf("memory contents changed\n");
+				errno = 33;
+			}
+			MemFree(arr[index]);
+			if (errno) {
+				errno = 0;
+				printf("run: %d, iteration: %d\n", r, i);
+			}
+			arr[index] = arr[ARR_SIZE - 1 - i];
+			sizes[index] = sizes[ARR_SIZE - 1 - i];
+		}
+	}
+	MemDone();
+	free(buf);
+	return 0;
+}
+
+/*int main() {
+	const int TEST_BLOCK_SIZE = 16;
+	const int TEST_MEMORY_SIZE = TEST_BLOCKS_COUNT * TEST_BLOCK_SIZE;
+	const int TEST_MEMORY_SIZE_INIT = 2 * TEST_BLOCKS_COUNT * (TEST_BLOCK_SIZE + MemGetBlockSize()); //multiplied by 2 so fragmentation will not fail the test
+	void* ptr = malloc(TEST_MEMORY_SIZE_INIT);
+	int bytes_init = MemInit(ptr, TEST_MEMORY_SIZE_INIT);
+	if (!bytes_init) {
+		free(ptr);
+		return -1;
+	}
+	void** blocks = (void**)malloc(TEST_BLOCKS_COUNT * sizeof(void*));
+	if (!blocks) {
+		free(ptr);
+		return -1;
+	}
+	for (int i = 0; i < TEST_BLOCKS_COUNT; i++) {
+		int size = rand() % (TEST_BLOCK_SIZE - 1) + 1;
+		blocks[i] = (char*)MemAlloc(size);
+		memset(blocks[i], 'a', size);
+	}
+	for (int i = 0; 2 * i < TEST_BLOCKS_COUNT; i++) {
+		MemFree(blocks[2 * i]);
+	}
+	for (int i = 0; 2 * i < TEST_BLOCKS_COUNT; i++) {
+		blocks[2 * i] = (char*)MemAlloc(TEST_BLOCK_SIZE);
+		memset(blocks[2 * i], 'a', TEST_BLOCK_SIZE);
+	}
+	for (int i = 0; 2 * i + 1 < TEST_BLOCKS_COUNT; i++) {
+		MemFree(blocks[2 * i + 1]);
+	}
+	for (int i = 0; 2 * i < TEST_BLOCKS_COUNT; i++) {
+		MemFree(blocks[2 * i]);
+	}
+	MemDone();
+	free(blocks);
+	free(ptr);
+	return 0;
+}*/
