+++ 3-Petroshenko-B15/LabB-Var15/to send.cc
@@ -1,5 +1,19 @@
-#include "LabB.h"
-#define TEST_BLOCKS_COUNT 6
+//FILE FOR TESTING SYSTEM
+
+#include <stdio.h>
+#include <stdlib.h>
+#include <memory.h>
+#include <math.h>
+#include "memallocator.h"
+
+typedef enum {
+	NoMerge,
+	MergeRight,
+	MergeLeft
+}Merge;
+
+#define Ok 1
+#define Error 0
 
 //pointer on the begining of the using heap
 void* sMemory = NULL;
@@ -26,14 +40,10 @@ int* GetFlagPtr(void* descriptor) {
 	return (int*)((char*)descriptor + *GetSizePtr(descriptor) - sizeof(int));
 }
 
-int GetAbs(int x) {
-	return x >= 0 ? x : -x;
-}
-
-int MemInit(void* pMemory, int size) {
+int meminit(void* pMemory, int size) {
 	if (!pMemory)
 		return Error;
-	if (size <= MemGetMinimumSize())
+	if (size <= memgetminimumsize())
 		return Error;
 	void* initdesc = pMemory;
 	*GetSizePtr(initdesc) = size;
@@ -46,7 +56,7 @@ int MemInit(void* pMemory, int size) {
 	return Ok;
 }
 
-void MemDone() {
+void memdone() {
 	void* iterator = sMemory;
 	while ((char*)iterator < (char*)sMemory + sizeofs)
 	{
@@ -58,18 +68,18 @@ void MemDone() {
 	return;
 }
 
-void* MemAlloc(int size) {
+void* memalloc(int size) {
 	if (size <= 0)
 		return NULL;
-	if (size > sizeofs - MemGetBlockSize())
+	if (size > sizeofs - memgetblocksize())
 		return NULL;
 	void* iterator = sHead;
 	void* fitblock = NULL;
 	int min = sizeofs + 1;
 	//searching for best fit block
-	while (iterator != NULL)
+	while (iterator)
 	{
-		if (*GetSizePtr(iterator) >= size + MemGetBlockSize() && *GetSizePtr(iterator) < min)
+		if (*GetSizePtr(iterator) >= size + memgetblocksize() && *GetSizePtr(iterator) < min)
 		{
 			fitblock = iterator;
 			min = *GetSizePtr(iterator);
@@ -80,11 +90,11 @@ void* MemAlloc(int size) {
 		return NULL;
 	//fragmentation:
 	//if we have enough memory to create new descriptor for empty part
-	if (*GetSizePtr(fitblock) > MemGetBlockSize() + size + MemGetBlockSize())
+	if (*GetSizePtr(fitblock) > memgetblocksize() + size + memgetblocksize())
 	{
 		//right part of the block, i.e. empty part
-		void* newdescriptor = (void*)((char*)fitblock + MemGetBlockSize() + size);
-		*GetSizePtr(newdescriptor) = *GetSizePtr(fitblock) - size - MemGetBlockSize();
+		void* newdescriptor = (void*)((char*)fitblock + memgetblocksize() + size);
+		*GetSizePtr(newdescriptor) = *GetSizePtr(fitblock) - size - memgetblocksize();
 		*GetNextPtr(newdescriptor) = *GetNextPtr(fitblock);
 		*GetPrevPtr(newdescriptor) = *GetPrevPtr(fitblock);
 		*GetFlagPtr(newdescriptor) = -*GetSizePtr(newdescriptor);
@@ -103,7 +113,7 @@ void* MemAlloc(int size) {
 				*GetPrevPtr(*GetNextPtr(fitblock)) = newdescriptor;
 			*GetNextPtr(*GetPrevPtr(fitblock)) = newdescriptor;
 		}
-		*GetSizePtr(fitblock) = MemGetBlockSize() + size;
+		*GetSizePtr(fitblock) = memgetblocksize() + size;
 	}
 	//if we give all block to user
 	else
@@ -111,28 +121,21 @@ void* MemAlloc(int size) {
 		//if this block is the first
 		if (!*GetPrevPtr(fitblock))
 		{
-			//if this block is not the only one in the list of empty blocks
-			if (*GetNextPtr(fitblock))
+			//if this block is the only one in the list of empty blocks
+			if (!*GetNextPtr(fitblock))
+				sHead = NULL;
+			else
 			{
 				*GetPrevPtr(*GetNextPtr(fitblock)) = NULL;
 				sHead = *GetNextPtr(fitblock);
 			}
-			else
-				sHead = NULL;
 		}
 		else
 		{
 			//if this block is not the last
 			if (*GetNextPtr(fitblock))
-			{
 				*GetPrevPtr(*GetNextPtr(fitblock)) = *GetPrevPtr(fitblock);
-				*GetNextPtr(*GetPrevPtr(fitblock)) = *GetNextPtr(fitblock);
-			}
-			else
-			{
-				*GetNextPtr(*GetPrevPtr(fitblock)) = NULL;
-				*GetPrevPtr(*GetNextPtr(fitblock)) = *GetPrevPtr(fitblock);
-			}
+			*GetNextPtr(*GetPrevPtr(fitblock)) = *GetNextPtr(fitblock);
 		}
 	}
 	//left part of the block, i.e. user's part
@@ -143,7 +146,7 @@ void* MemAlloc(int size) {
 	return (void*)((char*)fitblock + sizeof(int) + sizeof(void*) + sizeof(void*));
 }
 
-void MemFree(void* p) {
+void memfree(void* p) {
 	if (!p)
 		return;
 	void* pdescriptor = (void*)((char*)p - sizeof(int) - sizeof(void*) - sizeof(void*));
@@ -179,10 +182,21 @@ void MemFree(void* p) {
 			sHead = pdescriptor;
 		}
 		//repoint next and prev pointers to each other not to lose them after merge
-		if (*GetPrevPtr(rightblock))
-			*GetNextPtr(*GetPrevPtr(rightblock)) = *GetNextPtr(rightblock);
-		if (*GetNextPtr(rightblock))
-			*GetPrevPtr(*GetNextPtr(rightblock)) = *GetPrevPtr(rightblock);
+		//if right block is the first
+		if (!*GetPrevPtr(rightblock))
+		{
+			//if right block is not the only one
+			if (*GetNextPtr(rightblock))
+				*GetPrevPtr(*GetNextPtr(rightblock)) = *GetPrevPtr(rightblock); // = NULL
+			sHead = *GetNextPtr(rightblock);
+		}
+		else
+		{
+			//if right block is not the last
+			if (*GetNextPtr(rightblock))
+				*GetPrevPtr(*GetNextPtr(rightblock)) = *GetPrevPtr(rightblock);
+			*GetNextPtr(*GetPrevPtr(rightblock)) = *GetNextPtr(rightblock);// maybe NULL
+		}
 		*GetSizePtr(pdescriptor) += *GetSizePtr(rightblock);
 		*GetFlagPtr(pdescriptor) = -*GetSizePtr(pdescriptor);
 	}
@@ -199,49 +213,10 @@ void MemFree(void* p) {
 	return;
 }
 
-int MemGetMinimumSize() {
+int memgetminimumsize() {
 	return sizeof(int) + sizeof(void*) + sizeof(void*) + sizeof(int);
 }
 
-int MemGetBlockSize() {
+int memgetblocksize() {
 	return sizeof(int) + sizeof(void*) + sizeof(void*) + sizeof(int);
-}
-
-int main() {
-	/*const int TEST_BLOCK_SIZE = 16;
-	const int TEST_MEMORY_SIZE = TEST_BLOCKS_COUNT * TEST_BLOCK_SIZE;
-	const int TEST_MEMORY_SIZE_INIT = 2 * TEST_BLOCKS_COUNT * (TEST_BLOCK_SIZE + MemGetBlockSize()); //multiplied by 2 so fragmentation will not fail the test
-	void* ptr = malloc(TEST_MEMORY_SIZE_INIT);
-	int bytes_init = MemInit(ptr, TEST_MEMORY_SIZE_INIT);
-	if (!bytes_init) {
-		free(ptr);
-		return -1;
-	}
-	void** blocks = (void**)malloc(TEST_BLOCKS_COUNT * sizeof(void*));
-	if (!blocks) {
-		free(ptr);
-		return -1;
-	}
-	for (int i = 0; i < TEST_BLOCKS_COUNT; i++) {
-		int size = rand() % (TEST_BLOCK_SIZE - 1) + 1;
-		blocks[i] = (char*)MemAlloc(size);
-		memset(blocks[i], 'a', size);
-	}
-	for (int i = 0; 2 * i < TEST_BLOCKS_COUNT; i++) {
-		MemFree(blocks[2 * i]);
-	}
-	for (int i = 0; 2 * i < TEST_BLOCKS_COUNT; i++) {
-		blocks[2 * i] = (char*)MemAlloc(TEST_BLOCK_SIZE);
-		memset(blocks[i], 'a', TEST_BLOCK_SIZE);
-	}
-	for (int i = 0; 2 * i + 1 < TEST_BLOCKS_COUNT; i++) {
-		MemFree(blocks[2 * i + 1]);
-	}
-	for (int i = 0; 2 * i < TEST_BLOCKS_COUNT; i++) {
-		MemFree(blocks[2 * i]);
-	}
-	MemDone();
-	free(blocks);
-	free(ptr);*/
-	return 0;
 }
